feat(FN-1618): fix merge conflict errors
