--- trade-finance-manager-api/src/types/types-helper.ts
@@ -1,1 +1,0 @@
-export type ValuesOf<T> = T[keyof T];
